ForeignCards: remove card warning * upgrade green moving * collect view data

+++ app/src/main/java/inc/aminkinen/foreigncards/ui/TrainActivity.kt
@@ -3,6 +3,7 @@ package inc.aminkinen.foreigncards.ui
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.content.Intent
+import android.support.v7.app.AlertDialog
 import android.widget.Button
 import android.widget.EditText
 import android.widget.SeekBar
@@ -12,19 +13,26 @@ import inc.aminkinen.foreigncards.R
 import inc.aminkinen.foreigncards.database.DbProvider
 import inc.aminkinen.foreigncards.entities.Card
 import inc.aminkinen.foreigncards.entities.Settings
+import inc.aminkinen.foreigncards.entities.View
+import inc.aminkinen.foreigncards.entities.enums.GroupKind
 import inc.aminkinen.foreigncards.entities.enums.TrainMode
 import inc.aminkinen.foreigncards.logic.GroupManager
+import java.text.SimpleDateFormat
+import java.util.*
+
 
 class TrainActivity : AppCompatActivity() {
     private val _db : DbProvider = DbProvider.Instance
     private val _settings : Settings = _db.getSettings()
 
     private var _wasShowed : Boolean = false
+    private var _startViewingTime : Date = Date()
+    private var _currView : View? = null
 
     private var _groupIdMoveUp : Int? = null
     private var _groupIdMoveDown : Int? = null
 
-    private var _groupFilling : Map<Int, Int> = _db.getGroupsFilling(_settings.CurrentLanguage)
+    private var _groupFilling : Map<Int, Int> = HashMap()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -32,9 +40,12 @@ class TrainActivity : AppCompatActivity() {
 
         val texts = Texts(this)
 
+        loadGroupsFilling()
+
         val cards = Navigator(this, texts, _db.getCards(_settings.GroupIdForTraining, _settings.CurrentLanguage))
         cards.setup()
 
+        setupRemoving(cards)
         setupMoving(cards)
         setupShowing(cards, texts)
         setupUpdating(cards, texts)
@@ -44,17 +55,38 @@ class TrainActivity : AppCompatActivity() {
             Log.info("Close button was clicked")
             goToMain()
         }
-        val remove = findViewById<Button>(R.id.at_button_remove)
-        remove.setOnClickListener {
+    }
+
+    private fun setupRemoving(cards : Navigator) {
+        val builder = AlertDialog.Builder(this)
+        builder.setTitle("[TBD] Card removing")
+        builder.setMessage("Are you sure? The history'll be lost!")
+        builder.setPositiveButton("Yes") { dialog, arg ->
             Log.info("Remove button was clicked, remove ${cards.currCard}")
 
             _db.removeCard(cards.currCard.Id);
             cards.next()
         }
+
+        builder.setNegativeButton("No") { dialog, arg ->
+        }
+
+        builder.setCancelable(true)
+        builder.setOnCancelListener { dialog ->
+        }
+
+        val remove = findViewById<Button>(R.id.at_button_remove)
+        remove.setOnClickListener {
+            builder.show()
+        }
     }
 
     private fun setupMoving(cards : Navigator) {
-        _groupIdMoveUp = GroupManager.getUpGroup(cards.currCard.GroupId, _groupFilling)
+        if (GroupManager.getKind(cards.currCard.GroupId).upKind() == GroupKind.Green)
+            _groupIdMoveUp = _settings.GreenReceiveGroupId
+        else
+            _groupIdMoveUp = GroupManager.getUpGroup(cards.currCard.GroupId, _groupFilling)
+
         _groupIdMoveDown = GroupManager.getDownGroup(cards.currCard.GroupId, _groupFilling)
 
         setupUpMoving(cards, _groupIdMoveUp)
@@ -90,8 +122,9 @@ class TrainActivity : AppCompatActivity() {
         cards.currCard.GroupId = id
         _db.updateCard(cards.currCard)
 
-        // TODO: move to function
-        _groupFilling = _db.getGroupsFilling(_settings.CurrentLanguage)
+        _currView!!.MovingGroupId = id
+
+        loadGroupsFilling()
 
         cards.next()
     }
@@ -106,14 +139,29 @@ class TrainActivity : AppCompatActivity() {
                 currCard.Transc = texts.transc.text.toString()
                 currCard.Example = texts.example.text.toString()
             }
-            if (!texts.group.text.toString().isEmpty())
-                currCard.GroupId = Integer.parseInt(texts.group.text.toString())
+            if (!texts.group.text.toString().isEmpty()) {
+                val newGroup = Integer.parseInt(texts.group.text.toString())
+                if (currCard.GroupId != newGroup) {
+                    currCard.GroupId = newGroup
+                    _currView!!.MovingGroupId = newGroup
+                }
+            }
 
             Log.info("Update card: $currCard")
             _db.updateCard(currCard)
 
-            // TODO: move to function
-            _groupFilling = _db.getGroupsFilling(_settings.CurrentLanguage)
+            loadGroupsFilling()
+
+            cards.next()
+        }
+    }
+
+    private fun loadGroupsFilling() {
+        _groupFilling = _db.getGroupsFilling(_settings.CurrentLanguage)
+        if (_settings.GreenReceiveGroupId == -1 ||
+                _groupFilling.getOrElse(_settings.GreenReceiveGroupId) { 0 } >= GroupManager.getKindLimit(GroupKind.Green)) {
+            _settings.GreenReceiveGroupId = GroupManager.findLowestGroup(GroupKind.Green, _groupFilling)
+            _db.updateSettings(_settings)
         }
     }
 
@@ -132,6 +180,11 @@ class TrainActivity : AppCompatActivity() {
     }
 
     private fun onCurrCardChanged(cards : Navigator, texts: Texts) {
+        val lv = _currView
+        if (lv != null) {
+            addView(lv)
+        }
+
         val currCard = cards.currCard
         Log.info("Show new card: $currCard")
 
@@ -149,6 +202,18 @@ class TrainActivity : AppCompatActivity() {
         setupMoving(cards)
 
         _wasShowed = false
+        _startViewingTime = Date()
+        _currView = View(currCard.Id)
+    }
+
+    private fun addView(v: View) {
+        val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
+        val duration = Date().time - _startViewingTime.time
+
+        v.Time = dateFormat.format(Date())
+        v.Duration = duration / 1000.0
+
+        _db.addView(v)
     }
 
     private fun goToMain() {
